+++ Sources/SourceKitLSP/CMakeLists.txt
@@ -5,7 +5,9 @@ add_library(SourceKitLSP STATIC
   DocumentSnapshot+FromFileContents.swift
   IndexProgressManager.swift
   IndexStoreDB+MainFilesProvider.swift
+  LanguageServerType.swift
   LanguageService.swift
+  MessageHandlingDependencyTracker.swift
   Rename.swift
   ResponseError+Init.swift
   SourceKitIndexDelegate.swift
@@ -16,6 +18,7 @@ add_library(SourceKitLSP STATIC
   TestDiscovery.swift
   TextEdit+IsNoop.swift
   WorkDoneProgressManager.swift
+  WorkDoneProgressState.swift
   Workspace.swift
 )
 target_sources(SourceKitLSP PRIVATE
